connect: user to product

+++ services/product.service.js
@@ -4,7 +4,7 @@ const fileService = require("../services/file.service");
 const PRODUCTS_ON_PAGE = 12;
 
 class ProductService {
-  async createProduct(product, pictures) {
+  async createProduct(product, pictures, userId) {
     const fileName = fileService.saveFiles(pictures);
     const {
       title,
@@ -15,22 +15,20 @@ class ProductService {
       time_created,
       time_to_take,
       location,
-      phone,
       images,
       status,
     } = product;
     return db.query(
-      `INSERT INTO products (title, author, category_id, description, amount, time_created, time_to_take, location, phone, images, status) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) RETURNING *`,
+      `INSERT INTO products (title, author_id, category_id, description, amount, time_created, time_to_take, location, images, status) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) RETURNING *`,
       [
         title,
-        author,
+        userId,
         category_id,
         description,
         amount,
         time_created,
         time_to_take,
         location,
-        phone,
         fileName,
         status,
       ]
@@ -40,21 +38,24 @@ class ProductService {
     if (pageQuery === undefined) {
       if (searchQuery) {
         return db.query(
-          `SELECT p.id, p.title, p.author, p.description, p.amount, p.location, p.phone, p.images, p.status, p.time_created, p.time_to_take, 
-          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category
+          `SELECT p.id, p.title, p.description, p.amount, p.location, p.images, p.status, p.time_created, p.time_to_take, 
+          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category,
+          (SELECT json_build_object('id', a.id, 'name', a.name, 'surname', a.surname, 'company_name', a.company_name, 'email', a.email, 'phone', a.phone) FROM user_account a WHERE a.id = p.author_id) as author
           FROM products p WHERE LOWER(p.title) LIKE '%' || LOWER($1) || '%' ORDER BY p.id`,
           [searchQuery]
         );
       } else {
-        return db.query(`SELECT p.id, p.title, p.author, p.description, p.amount, p.location, p.phone, p.images, p.status, p.time_created, p.time_to_take, 
-        (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category
+        return db.query(`SELECT p.id, p.title, p.description, p.amount, p.location, p.images, p.status, p.time_created, p.time_to_take, 
+        (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category,
+        (SELECT json_build_object('id', a.id, 'name', a.name, 'surname', a.surname, 'company_name', a.company_name, 'email', a.email, 'phone', a.phone) FROM user_account a WHERE a.id = p.author_id) as author
         FROM products p ORDER BY p.id`);
       }
     } else if (Number(pageQuery)) {
       if (searchQuery) {
         return db.query(
-          `SELECT p.id, p.title, p.author, p.description, p.amount, p.location, p.phone, p.images, p.status, p.time_created, p.time_to_take, 
-          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category
+          `SELECT p.id, p.title, p.description, p.amount, p.location, p.images, p.status, p.time_created, p.time_to_take, 
+          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category,
+          (SELECT json_build_object('id', a.id, 'name', a.name, 'surname', a.surname, 'company_name', a.company_name, 'email', a.email, 'phone', a.phone) FROM user_account a WHERE a.id = p.author_id) as author
           FROM products p WHERE LOWER(p.title) LIKE '%' || LOWER($1) || '%' ORDER BY p.id LIMIT ${PRODUCTS_ON_PAGE} OFFSET ${
             PRODUCTS_ON_PAGE * (pageQuery - 1)
           }`,
@@ -62,8 +63,9 @@ class ProductService {
         );
       } else {
         return db.query(
-          `SELECT p.id, p.title, p.author, p.description, p.amount, p.location, p.phone, p.images, p.status, p.time_created, p.time_to_take, 
-          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category
+          `SELECT p.id, p.title, p.description, p.amount, p.location, p.images, p.status, p.time_created, p.time_to_take, 
+          (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category,
+          (SELECT json_build_object('id', a.id, 'name', a.name, 'surname', a.surname, 'company_name', a.company_name, 'email', a.email, 'phone', a.phone) FROM user_account a WHERE a.id = p.author_id) as author
           FROM products p ORDER BY p.id LIMIT ${PRODUCTS_ON_PAGE} OFFSET ${
             PRODUCTS_ON_PAGE * (pageQuery - 1)
           }`
@@ -83,15 +85,15 @@ class ProductService {
   }
   async getOneProduct(productId) {
     return db.query(
-      `SELECT p.id, p.title, p.author, p.description, p.amount, p.location, p.phone, p.images, p.status, p.time_created, p.time_to_take, 
-      (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category
+      `SELECT p.id, p.title, p.description, p.amount, p.location, p.images, p.status, p.time_created, p.time_to_take, 
+      (SELECT json_build_object('id', c.id, 'name', c.name) FROM category c WHERE c.id = p.category_id) as category,
+      (SELECT json_build_object('id', a.id, 'name', a.name, 'surname', a.surname, 'company_name', a.company_name, 'email', a.email, 'phone', a.phone) FROM user_account a WHERE a.id = p.author_id) as author
       FROM products p WHERE p.id = $1`,
       [productId]
     );
   }
 
-  async updateProduct(product, pictures) {
-    const fileName = fileService.saveFiles(pictures);
+  async updateProduct(product, pictures, userId) {
     const {
       title,
       author,
@@ -101,35 +103,55 @@ class ProductService {
       time_created,
       time_to_take,
       location,
-      phone,
       images,
       status,
       id,
     } = product;
-    return db.query(
-      `UPDATE products SET title = $1, author = $2, category_id = $3, description = $4, amount = $5, time_created = $6, time_to_take = $7, location = $8, phone = $9, images = $10, status = $11 where id = $12 RETURNING *`,
-      [
-        title,
-        author,
-        category_id,
-        description,
-        amount,
-        time_created,
-        time_to_take,
-        location,
-        phone,
-        fileName,
-        status,
-        id,
-      ]
-    );
+    if (Number(author) !== userId) {
+      console.log("ERROR AUTH");
+    } else {
+      // delete pictures before adding new
+      let picturesToDelete = await db.query(
+        `SELECT images FROM products where id = $1`,
+        [id]
+      );
+      fileService.deleteFiles(picturesToDelete[0].images);
+
+      // add new pictures
+      const fileName = fileService.saveFiles(pictures);
+      return db.query(
+        `UPDATE products SET title = $1, author_id = $2, category_id = $3, description = $4, amount = $5, time_created = $6, time_to_take = $7, location = $8, images = $9, status = $10 where id = $11 RETURNING *`,
+        [
+          title,
+          userId,
+          category_id,
+          description,
+          amount,
+          time_created,
+          time_to_take,
+          location,
+          fileName,
+          status,
+          id,
+        ]
+      );
+    }
   }
-  async deleteProduct(productId) {
-    let pictures = await db.query(`SELECT images FROM products where id = $1`, [
-      productId,
-    ]);
-    fileService.deleteFiles(pictures[0].images);
-    return db.query(`DELETE FROM products where id = $1`, [productId]);
+  async deleteProduct(productId, userId) {
+    let productData = await db.query(
+      `SELECT images, author_id FROM products where id = $1`,
+      [productId]
+    );
+    if (Number(productData[0].author_id) !== userId) {
+      console.log("ERROR AUTH");
+    } else {
+      let pictures = await db.query(
+        `SELECT images FROM products where id = $1`,
+        [productId]
+      );
+      fileService.deleteFiles(pictures[0].images);
+      return db.query(`DELETE FROM products where id = $1`, [productId]);
+    }
   }
 }
